app/self-hosted-models/whisper-transcription/requirements.txt
fastapi==0.95.0
pydantic==1.10.6
uvicorn[standard]
whisper-openai
requests=2.28.2